Honor custom name when TM is registered with the legacy mode. (#38547)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38547

Before this change, when a turbomodule was registered using the `RCT_EXPORT_MODULE()` function using a custom name, the TM was not found in the registry.

With this change, we are adding a fallback that looks in the list of registered TM, asking for the name used by the user to register it.

This fallback is executed only if no other method was successful before as it is more expensive.

## CHANGELOG
[iOS][Fixed] - Honor the custom name choosen by the user to register the module.

Reviewed By: philIip

Differential Revision: D47666848

fbshipit-source-id: 545855de1c398726b9449a0d77af7f1756cca233