Update facebook::react::kNumberOfPropsPerComponentSoftCap
Summary:
`butter::small_vector` is backed by a `folly:small_vector` on non-Android platforms in release mode.

It uses the small vector optimization, keeping a fixed size buffer, then falling back to heap allocation.

When I was debugging through props parsing code to figure out a prop not getting passed to Fabric (was being filtered in ViewConfig), I noticed this vector is filled with 199 elements for ViewProps. So It always overflows the current capacity and falls back to heap allocation. This bumps the capacity basde on what I observed with some more headroom.

Changelog:
[Internal]

Reviewed By: sammy-SC

Differential Revision: D47981105

fbshipit-source-id: 3870d8e4fee9748c487d01cc82284de865ea370d